Inline ckbtc_minter_retrieve_btc
