+++ rs/tests/src/ckbtc/minter/mod.rs
@@ -1,3 +1,2 @@
-pub mod test_retrieve_btc;
 pub mod test_update_balance;
 pub mod utils;
